# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proq API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Create an account and, with it, the organization that owns its data.\n\nEvery signup here gets its own new tenant. Joining an EXISTING organization\ngoes through the invite flow instead (POST /api/invite/{token}/accept in\napp/api/routes/team.py), which creates the user in the inviting org.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/auth/email-config": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Email Config",
        "description": "The effective outbound-email setup: which mailbox mail leaves from,\nwhether Gmail is actually connected, and your Cc address.\n\nEvery field derives from the PROCUREAI_GMAIL_* environment variables (see\ndocs/email-setup.md) \u2014 nothing here is per-user except `ccEmail` and the\ndisplay name baked into `fromHeader`.",
        "operationId": "email_config_api_auth_email_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailConfig"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/auth/test-email": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Test Email",
        "description": "Verify the email configuration by sending a test message to yourself.\n\nUses exactly the same path as an RFQ send: the configured provider (Gmail\nor the logging mock) and the workspace From address carrying your display\nname. See docs/email-setup.md.",
        "operationId": "send_test_email_api_auth_test_email_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmailResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailConfig": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "mocked": {
            "type": "boolean",
            "title": "Mocked"
          },
          "senderAddressSet": {
            "type": "boolean",
            "title": "Senderaddressset"
          },
          "fromAddress": {
            "type": "string",
            "title": "Fromaddress"
          },
          "fromHeader": {
            "type": "string",
            "title": "Fromheader"
          },
          "ccEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ccemail"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing",
            "default": []
          },
          "gmail": {
            "type": "object",
            "title": "Gmail",
            "default": {}
          },
          "llm": {
            "type": "object",
            "title": "Llm",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "configured",
          "mocked",
          "senderAddressSet",
          "fromAddress",
          "fromHeader"
        ],
        "title": "EmailConfig",
        "description": "Effective outbound-email configuration, straight from the environment.\n\nLets the UI state the truth instead of implying mail is going out: when\n`configured` is false nothing is delivered, and when `senderAddressSet` is\nfalse `fromAddress` is only a placeholder."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "company": {
            "type": "string",
            "title": "Company",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "TestEmailResult": {
        "properties": {
          "mocked": {
            "type": "boolean",
            "title": "Mocked"
          },
          "messageId": {
            "type": "string",
            "title": "Messageid"
          },
          "fromAddr": {
            "type": "string",
            "title": "Fromaddr"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "cc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc"
          }
        },
        "type": "object",
        "required": [
          "mocked",
          "messageId",
          "fromAddr",
          "to"
        ],
        "title": "TestEmailResult",
        "description": "Outcome of POST /api/auth/test-email (config verification)."
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "UpdateMeRequest": {
        "properties": {
          "ccEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ccemail"
          }
        },
        "type": "object",
        "title": "UpdateMeRequest",
        "description": "Editable account settings. `ccEmail` is the address copied on outgoing\nmail you trigger; `null` clears it. It is never a From address \u2014 everything\nis sent from the workspace mailbox (see services/rfq/sender.py)."
      },
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "organizationId": {
            "type": "string",
            "title": "Organizationid"
          },
          "ccEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ccemail"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "company",
          "organizationId"
        ],
        "title": "User",
        "description": "Public-safe user shape (never includes the password hash)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}